pyproject.toml
[project]
name = "sms-wallet-demo"
version = "0.1.0"
description = "Demo-grade SMS wallet platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.0",
  "jinja2>=3.1.0",
  "passlib[bcrypt]>=1.7.4",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
test = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
